requirements.txt
Django>=5.1.2,<5.2.0
djangorestframework>=3.15.2,<3.16.0
drf-spectacular>=0.27.2,<0.28.0